add rating feature

+++ src/main/java/com/example/popcorn_films/entity/Rating.java
@@ -5,12 +5,10 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
@@ -34,7 +32,7 @@ public class Rating {
     @Column(nullable = false)
     @Min(1)
     @Max(10)
-    private Integer mark;
+    private Long mark;
 
     @ManyToOne
     private User user;
